Use unsigned long to store pid

+++ src/misc.hpp
@@ -4,6 +4,6 @@
 
 std::string utf16_to_utf8(const std::wstring& utf16_string);
 
-std::string GetProcessExecName(const unsigned int pid);
+std::string GetProcessExecName(const unsigned long pid);
 
 std::string GetLastErrorMessage();
